Issue with Decimal Datatype in API

+++ eFacilito_MobileApp_WebAPI/eFacilito_MobileApp_WebAPI/Controllers/TicketingController.cs
@@ -123,7 +123,7 @@ namespace eFacilito_MobileApp_WebAPI.Controllers
                                              SubCategory_Desc = p.Field<string>("SubCategory_Desc"),
                                              Ticket_Date = p.Field<string>("Ticket_Date"),
                                              Ticket_Status = p.Field<string>("Tkt_Status"),
-                                             Level = Convert.ToString(p.Field<decimal>("Tkt_Level")),
+                                             Level = Convert.ToString(p.Field<Int32>("Tkt_Level")),
                                              Ticket_ActionStatus = p.Field<string>("Tkt_ActionStatus"),
                                              Ticket_Message = p.Field<string>("Tkt_Message"),
                                              Ticket_ImagePath = p.Field<string>("ImagePath")
@@ -202,7 +202,7 @@ namespace eFacilito_MobileApp_WebAPI.Controllers
                                              SubCategory_Desc = p.Field<string>("SubCategory_Desc"),
                                              Ticket_Date = p.Field<string>("Ticket_Date"),
                                              Ticket_Status = p.Field<string>("Tkt_Status"),
-                                             Level = Convert.ToString(p.Field<decimal>("Tkt_Level")),
+                                             Level = Convert.ToString(p.Field<Int32>("Tkt_Level")),
                                              Ticket_ActionStatus = p.Field<string>("Tkt_ActionStatus"),
                                              Ticket_Message = p.Field<string>("Tkt_Message"),
                                              Ticket_ImagePath = p.Field<string>("ImagePath")
